backend/pyproject.toml
[project]
name = "jorg-backend"
version = "0.1.0"
description = "Backend for Jorg - consulting skill profile platform"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "authlib>=1.3.2",
    "httpx>=0.27",
    "email-validator>=2.2",
    "structlog>=24.4",
    "python-docx>=1.1",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "testcontainers[postgres]>=4.8",
    "ruff>=0.6",
    "mypy>=1.11",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["E501"]
"tests/*" = ["E501"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["jose", "jose.exceptions", "testcontainers.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "tests.unit.*", "tests.integration.*", "alembic.*", "alembic.versions.*"]
disallow_untyped_defs = false
disallow_any_generics = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "alembic/*"]
